Don't lazily-evaluate the Passenger root.
There doesn't seem to be a good way to do this in current versions of
Chef.  Here's the problem I'm hitting:

1. gem_package installs Passenger.  Unless I specify this elsewhere, I
   don't know what version was installed, and as such must use
   passenger-config --root.
2. I have to be able to propagate attributes derived from
   passenger-config --root to template resources.  However, the
   template values have been fixed in the Chef compile phase.
3. There is no way to reload a template resource.
4. There is no way (well, as far as I can tell) to lazily evaluate
   attributes.

The solution: eagerly evaluate everything.  This does have some
unpleasant surprises -- for example, passenger[:version] will not accept
Rubygems' {>, >=, <, <=, ~>} version constraints -- but it does have the
nice effects of

1. working
2. letting us fix a Passenger version for all nodes