The bcdatabase key file needs the same permissions as the group files.
If a user is permitted to read a group file because they need to use
the credentials in that group file, that permission is useless in the
presence of encrypted passwords.  The key must also be accessible to
said user.

I do realize that making group and key files world-readable pretty much
negates the point of even having a key.  (The key also doesn't do
anything against a smart attacker anyway.)  I'll revisit the usage of
said key in future commits.