Cookbook for configuring zeroconf via avahi in Vagrant.
Why?

The problem:

Section 2.5.4 of the CAS 2.0 specification requires that a CAS proxy
callback (1) be accessible over HTTPS and (2) present a certificate that
the server can verify as valid.  The CAS server uses the verification
logic provided by the Java class library.

It is possible to use IP addresses in SSL certificates.  RFC 2818 (which
describes the most widely-used HTTPS protocol) prescribes the following
procedure for verification:

    In some cases, the URI is specified as an IP address rather than a
    hostname.  In this case, the iPAddress subjectAltName must be
    present in the certificate and must exactly match the IP in the URI.

The verification code in the Java class library seems to follow this
logic.

However, while all of this works, it feels very clumsy in practice.
Most SSL/TLS-using systems work best when they're comparing names; RFC
6125 (which generalizes TLS certificate verification beyond HTTPS)
describes it as "out of scope", which is a pretty good indication that
support for verification-by-IP will slowly continue to fall out of
favor.

Fortunately, it is not difficult to generate a name for our Vagrant
nodes.  Zeroconf can do that for us, making the host-only interface of
the VM resolvable as [hostname].local from hosts and guests.