Bring application_users back.
This is part of my Plan B to sort out permissions on the CAS node: run
the CAS server and the callback as user cas.  In an environment running
multiple applications, this would be pretty nasty, but it should work
fine for our usual way in which the CAS server gets its own node.