More lenient permissions.
There's nothing sensitive in these logs[0], and having to elevate
privileges just to read them is a security hazard.

[0] Well, almost.  The CAS log, at present, contains ticket texts, so
there exists a small window in which an attacker could intercept service
or proxy tickets for their own use.  This can probably be fixed by
changing the CAS server's log level.