Don't overwrite role defaults for CAS' SSL certificates.
This causes problems on first runs when using a role that supplies
default attributes for SSL certificates: said attributes can be
overwritten by the defaults specified in the CAS cookbook.

Another way to fix this would be to have the attribute-supplying role
use overrides (and I think I'll do that anyway), but the only purpose of
the nil defaults was for development environment setup, and on second
thought I think assigning a Mash to cas.apache.ssl is a cleaner way to
go.