Fix paths sent to the ssl_certificate LWRP; clarify LWRP actions.
Prior to this commit, the ssl_certificates LWRP was receiving
basenames of certificates, not absolute paths.  This caused a silent
failure (!) of the symlinking script.

To do: make the symlinking script cause a Chef run to abort if it fails.
Bonus points: find a way to make use of the link resource.