Immediately rebuild the trust chain.
If the trust chain bundle isn't rebuilt before a process using that
bundle is (re)started, we can end up in a situation where said process
uses an old copy of that bundle on restart.  To prevent this from
happening, we need to properly sequence restarts and rebuilds, and the
easiest way to do that is to rebuild the bundle as soon as the need for
a rebuild is detected.